refactor: adjust naming convention
